Moved written files to the right folders
