--- src/main/java/de/uni_stuttgart/it_rex/course/repository/CourseRepository.java
@@ -1,14 +1,0 @@
-package de.uni_stuttgart.it_rex.course.repository;
-
-import de.uni_stuttgart.it_rex.course.domain.Course;
-
-import org.springframework.data.jpa.repository.*;
-import org.springframework.stereotype.Repository;
-
-/**
- * Spring Data  repository for the Course entity.
- */
-@SuppressWarnings("unused")
-@Repository
-public interface CourseRepository extends JpaRepository<Course, Long> {
-}
