+++ src/main/java/de/uni_stuttgart/it_rex/course/repository/written/ParticipationRepository.java
@@ -1,0 +1,20 @@
+package de.uni_stuttgart.it_rex.course.repository.written;
+
+import de.uni_stuttgart.it_rex.course.domain.written.Participation;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.stereotype.Repository;
+
+import java.util.List;
+import java.util.UUID;
+
+/**
+ * Spring Data  repository for the Participation entity.
+ */
+@SuppressWarnings("unused")
+@Repository
+public interface ParticipationRepository extends JpaRepository<Participation, UUID> {
+
+    @Query("select participation from Participation participation where participation.user.login = ?#{principal.preferredUsername}")
+    List<Participation> findByUserIsCurrentUser();
+}
