+++ src/main/java/de/uni_stuttgart/it_rex/course/service/written/CourseService.java
@@ -1,0 +1,86 @@
+package de.uni_stuttgart.it_rex.course.service.written;
+
+import de.uni_stuttgart.it_rex.course.domain.written.Course;
+import de.uni_stuttgart.it_rex.course.repository.written.CourseRepository;
+import de.uni_stuttgart.it_rex.course.service.written.dto.CourseDTO;
+import de.uni_stuttgart.it_rex.course.service.written.mapper.CourseMapper;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.util.LinkedList;
+import java.util.List;
+import java.util.Optional;
+import java.util.UUID;
+import java.util.stream.Collectors;
+
+/**
+ * Service Implementation for managing {@link Course}.
+ */
+@Service
+@Transactional
+public class CourseService {
+
+    private final Logger log = LoggerFactory.getLogger(CourseService.class);
+
+    private final CourseRepository courseRepository;
+
+    private final CourseMapper courseMapper;
+
+    public CourseService(CourseRepository courseRepository, CourseMapper courseMapper) {
+        this.courseRepository = courseRepository;
+        this.courseMapper = courseMapper;
+    }
+
+    /**
+     * Save a course.
+     *
+     * @param courseDTO the entity to save.
+     * @return the persisted entity.
+     */
+    public CourseDTO save(CourseDTO courseDTO) {
+        log.debug("Request to save Course : {}", courseDTO);
+        Course course = courseMapper.toEntity(courseDTO);
+        course = courseRepository.save(course);
+        return courseMapper.toDto(course);
+    }
+
+    /**
+     * Get all the courses.
+     *
+     * @return the list of entities.
+     */
+    @Transactional(readOnly = true)
+    public List<CourseDTO> findAll() {
+        log.debug("Request to get all Courses");
+        return courseRepository.findAll().stream()
+            .map(courseMapper::toDto)
+            .collect(Collectors.toCollection(LinkedList::new));
+    }
+
+
+    /**
+     * Get one course by id.
+     *
+     * @param id the id of the entity.
+     * @return the entity.
+     */
+    @Transactional(readOnly = true)
+    public Optional<CourseDTO> findOne(final UUID id) {
+        log.debug("Request to get Course : {}", id);
+        return courseRepository.findById(id)
+            .map(courseMapper::toDto);
+    }
+
+    /**
+     * Delete the course by id.
+     *
+     * @param id the id of the entity.
+     */
+    public void delete(final UUID id) {
+        log.debug("Request to delete Course : {}", id);
+        courseRepository.deleteById(id);
+    }
+}
