+++ src/main/java/de/uni_stuttgart/it_rex/course/service/written/mapper/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * Custom MapStruct mappers for mapping domain objects and Data Transfer
+ * Objects.
+ */
+package de.uni_stuttgart.it_rex.course.service.written.mapper;
