+++ src/main/java/de/uni_stuttgart/it_rex/course/service/written/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Custom service layer beans.
+ */
+package de.uni_stuttgart.it_rex.course.service.written;
