--- src/main/java/de/uni_stuttgart/it_rex/course/web/rest/CourseResource.java
@@ -1,116 +1,0 @@
-package de.uni_stuttgart.it_rex.course.web.rest;
-
-import de.uni_stuttgart.it_rex.course.service.CourseService;
-import de.uni_stuttgart.it_rex.course.web.rest.errors.BadRequestAlertException;
-import de.uni_stuttgart.it_rex.course.service.dto.CourseDTO;
-
-import io.github.jhipster.web.util.HeaderUtil;
-import io.github.jhipster.web.util.ResponseUtil;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-
-import java.net.URI;
-import java.net.URISyntaxException;
-import java.util.List;
-import java.util.Optional;
-
-/**
- * REST controller for managing {@link de.uni_stuttgart.it_rex.course.domain.Course}.
- */
-@RestController
-@RequestMapping("/api")
-public class CourseResource {
-
-    private final Logger log = LoggerFactory.getLogger(CourseResource.class);
-
-    private static final String ENTITY_NAME = "courseServiceCourse";
-
-    @Value("${jhipster.clientApp.name}")
-    private String applicationName;
-
-    private final CourseService courseService;
-
-    public CourseResource(CourseService courseService) {
-        this.courseService = courseService;
-    }
-
-    /**
-     * {@code POST  /courses} : Create a new course.
-     *
-     * @param courseDTO the courseDTO to create.
-     * @return the {@link ResponseEntity} with status {@code 201 (Created)} and with body the new courseDTO, or with status {@code 400 (Bad Request)} if the course has already an ID.
-     * @throws URISyntaxException if the Location URI syntax is incorrect.
-     */
-    @PostMapping("/courses")
-    public ResponseEntity<CourseDTO> createCourse(@RequestBody CourseDTO courseDTO) throws URISyntaxException {
-        log.debug("REST request to save Course : {}", courseDTO);
-        if (courseDTO.getId() != null) {
-            throw new BadRequestAlertException("A new course cannot already have an ID", ENTITY_NAME, "idexists");
-        }
-        CourseDTO result = courseService.save(courseDTO);
-        return ResponseEntity.created(new URI("/api/courses/" + result.getId()))
-            .headers(HeaderUtil.createEntityCreationAlert(applicationName, true, ENTITY_NAME, result.getId().toString()))
-            .body(result);
-    }
-
-    /**
-     * {@code PUT  /courses} : Updates an existing course.
-     *
-     * @param courseDTO the courseDTO to update.
-     * @return the {@link ResponseEntity} with status {@code 200 (OK)} and with body the updated courseDTO,
-     * or with status {@code 400 (Bad Request)} if the courseDTO is not valid,
-     * or with status {@code 500 (Internal Server Error)} if the courseDTO couldn't be updated.
-     * @throws URISyntaxException if the Location URI syntax is incorrect.
-     */
-    @PutMapping("/courses")
-    public ResponseEntity<CourseDTO> updateCourse(@RequestBody CourseDTO courseDTO) throws URISyntaxException {
-        log.debug("REST request to update Course : {}", courseDTO);
-        if (courseDTO.getId() == null) {
-            throw new BadRequestAlertException("Invalid id", ENTITY_NAME, "idnull");
-        }
-        CourseDTO result = courseService.save(courseDTO);
-        return ResponseEntity.ok()
-            .headers(HeaderUtil.createEntityUpdateAlert(applicationName, true, ENTITY_NAME, courseDTO.getId().toString()))
-            .body(result);
-    }
-
-    /**
-     * {@code GET  /courses} : get all the courses.
-     *
-     * @return the {@link ResponseEntity} with status {@code 200 (OK)} and the list of courses in body.
-     */
-    @GetMapping("/courses")
-    public List<CourseDTO> getAllCourses() {
-        log.debug("REST request to get all Courses");
-        return courseService.findAll();
-    }
-
-    /**
-     * {@code GET  /courses/:id} : get the "id" course.
-     *
-     * @param id the id of the courseDTO to retrieve.
-     * @return the {@link ResponseEntity} with status {@code 200 (OK)} and with body the courseDTO, or with status {@code 404 (Not Found)}.
-     */
-    @GetMapping("/courses/{id}")
-    public ResponseEntity<CourseDTO> getCourse(@PathVariable Long id) {
-        log.debug("REST request to get Course : {}", id);
-        Optional<CourseDTO> courseDTO = courseService.findOne(id);
-        return ResponseUtil.wrapOrNotFound(courseDTO);
-    }
-
-    /**
-     * {@code DELETE  /courses/:id} : delete the "id" course.
-     *
-     * @param id the id of the courseDTO to delete.
-     * @return the {@link ResponseEntity} with status {@code 204 (NO_CONTENT)}.
-     */
-    @DeleteMapping("/courses/{id}")
-    public ResponseEntity<Void> deleteCourse(@PathVariable Long id) {
-        log.debug("REST request to delete Course : {}", id);
-        courseService.delete(id);
-        return ResponseEntity.noContent().headers(HeaderUtil.createEntityDeletionAlert(applicationName, true, ENTITY_NAME, id.toString())).build();
-    }
-}
