+++ src/main/java/de/uni_stuttgart/it_rex/course/web/rest/written/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Custom Spring MVC REST controllers.
+ */
+package de.uni_stuttgart.it_rex.course.web.rest.written;
