--- src/main/java/de/uni_stuttgart/it_rex/course/written/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Custom implementations.
- */
-package de.uni_stuttgart.it_rex.course.written;
