--- src/main/java/de/uni_stuttgart/it_rex/course/written/service/mapper/package-info.java
@@ -1,5 +1,0 @@
-/**
- * Custom MapStruct mappers for mapping domain objects and Data Transfer
- * Objects.
- */
-package de.uni_stuttgart.it_rex.course.written.service.mapper;
