--- src/main/java/de/uni_stuttgart/it_rex/course/written/service/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Custom service layer beans.
- */
-package de.uni_stuttgart.it_rex.course.written.service;
