--- src/main/java/de/uni_stuttgart/it_rex/course/written/web/rest/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Custom Spring MVC REST controllers.
- */
-package de.uni_stuttgart.it_rex.course.written.web.rest;
