+++ src/main/resources/config/liquibase/changelog/20210113223125_added_entity_Course.xml
@@ -1,22 +1,17 @@
 <?xml version="1.0" encoding="utf-8"?>
 <databaseChangeLog
-    xmlns="http://www.liquibase.org/xml/ns/dbchangelog"
-    xmlns:ext="http://www.liquibase.org/xml/ns/dbchangelog-ext"
-    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-    xsi:schemaLocation="http://www.liquibase.org/xml/ns/dbchangelog http://www.liquibase.org/xml/ns/dbchangelog/dbchangelog-3.9.xsd
-                        http://www.liquibase.org/xml/ns/dbchangelog-ext http://www.liquibase.org/xml/ns/dbchangelog/dbchangelog-ext.xsd">
+        xmlns="http://www.liquibase.org/xml/ns/dbchangelog"
+        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+        xsi:schemaLocation="http://www.liquibase.org/xml/ns/dbchangelog http://www.liquibase.org/xml/ns/dbchangelog/dbchangelog-3.9.xsd">
 
-    
+
     <!--
         Added the entity Course.
     -->
     <changeSet id="20210113223125-1" author="jhipster">
         <createTable tableName="course">
-            <column name="id" type="bigint">
+            <column name="id" type="${uuidType}">
                 <constraints primaryKey="true" nullable="false"/>
-            </column>
-            <column name="uuid" type="${uuidType}">
-                <constraints nullable="true" />
             </column>
             <column name="name" type="varchar(255)">
                 <constraints nullable="true" />
