--- src/test/java/de/uni_stuttgart/it_rex/course/service/dto/CourseDTOTest.java
@@ -1,23 +1,0 @@
-package de.uni_stuttgart.it_rex.course.service.dto;
-
-import org.junit.jupiter.api.Test;
-import static org.assertj.core.api.Assertions.assertThat;
-import de.uni_stuttgart.it_rex.course.web.rest.TestUtil;
-
-public class CourseDTOTest {
-
-    @Test
-    public void dtoEqualsVerifier() throws Exception {
-        TestUtil.equalsVerifier(CourseDTO.class);
-        CourseDTO courseDTO1 = new CourseDTO();
-        courseDTO1.setId(1L);
-        CourseDTO courseDTO2 = new CourseDTO();
-        assertThat(courseDTO1).isNotEqualTo(courseDTO2);
-        courseDTO2.setId(courseDTO1.getId());
-        assertThat(courseDTO1).isEqualTo(courseDTO2);
-        courseDTO2.setId(2L);
-        assertThat(courseDTO1).isNotEqualTo(courseDTO2);
-        courseDTO1.setId(null);
-        assertThat(courseDTO1).isNotEqualTo(courseDTO2);
-    }
-}
