--- src/test/java/de/uni_stuttgart/it_rex/course/service/mapper/CourseMapperTest.java
@@ -1,22 +1,0 @@
-package de.uni_stuttgart.it_rex.course.service.mapper;
-
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-import static org.assertj.core.api.Assertions.assertThat;
-
-public class CourseMapperTest {
-
-    private CourseMapper courseMapper;
-
-    @BeforeEach
-    public void setUp() {
-        courseMapper = new CourseMapperImpl();
-    }
-
-    @Test
-    public void testEntityFromId() {
-        Long id = 1L;
-        assertThat(courseMapper.fromId(id).getId()).isEqualTo(id);
-        assertThat(courseMapper.fromId(null)).isNull();
-    }
-}
