+++ src/test/java/de/uni_stuttgart/it_rex/course/service/mapper/written/dto/CourseDTOTest.java
@@ -1,0 +1,29 @@
+package de.uni_stuttgart.it_rex.course.service.mapper.written.dto;
+
+import de.uni_stuttgart.it_rex.course.service.written.dto.CourseDTO;
+import de.uni_stuttgart.it_rex.course.web.rest.TestUtil;
+import org.junit.jupiter.api.Test;
+
+import java.util.UUID;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+public class CourseDTOTest {
+    private static final UUID FIRST_ID = UUID.randomUUID();
+    private static final UUID SECOND_ID = UUID.randomUUID();
+
+    @Test
+    public void dtoEqualsVerifier() throws Exception {
+        TestUtil.equalsVerifier(CourseDTO.class);
+        CourseDTO courseDTO1 = new CourseDTO();
+        courseDTO1.setId(FIRST_ID);
+        CourseDTO courseDTO2 = new CourseDTO();
+        assertThat(courseDTO1).isNotEqualTo(courseDTO2);
+        courseDTO2.setId(courseDTO1.getId());
+        assertThat(courseDTO1).isEqualTo(courseDTO2);
+        courseDTO2.setId(SECOND_ID);
+        assertThat(courseDTO1).isNotEqualTo(courseDTO2);
+        courseDTO1.setId(null);
+        assertThat(courseDTO1).isNotEqualTo(courseDTO2);
+    }
+}
