--- src/test/java/de/uni_stuttgart/it_rex/course/web/rest/ParticipationResourceIT.java
@@ -1,235 +1,0 @@
-package de.uni_stuttgart.it_rex.course.web.rest;
-
-import de.uni_stuttgart.it_rex.course.CourseServiceApp;
-import de.uni_stuttgart.it_rex.course.config.TestSecurityConfiguration;
-import de.uni_stuttgart.it_rex.course.domain.Participation;
-import de.uni_stuttgart.it_rex.course.repository.ParticipationRepository;
-import de.uni_stuttgart.it_rex.course.service.ParticipationService;
-import de.uni_stuttgart.it_rex.course.service.dto.ParticipationDTO;
-import de.uni_stuttgart.it_rex.course.service.mapper.ParticipationMapper;
-
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.http.MediaType;
-import org.springframework.security.test.context.support.WithMockUser;
-import org.springframework.test.web.servlet.MockMvc;
-import org.springframework.transaction.annotation.Transactional;
-import javax.persistence.EntityManager;
-import java.util.List;
-import java.util.UUID;
-
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.hamcrest.Matchers.hasItem;
-import static org.springframework.security.test.web.servlet.request.SecurityMockMvcRequestPostProcessors.csrf;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
-
-import de.uni_stuttgart.it_rex.course.domain.enumeration.PARTICIPATIONTYPE;
-/**
- * Integration tests for the {@link ParticipationResource} REST controller.
- */
-@SpringBootTest(classes = { CourseServiceApp.class, TestSecurityConfiguration.class })
-@AutoConfigureMockMvc
-@WithMockUser
-public class ParticipationResourceIT {
-
-    private static final UUID DEFAULT_UUID = UUID.randomUUID();
-    private static final UUID UPDATED_UUID = UUID.randomUUID();
-
-    private static final PARTICIPATIONTYPE DEFAULT_TYPE = PARTICIPATIONTYPE.STUDENT;
-    private static final PARTICIPATIONTYPE UPDATED_TYPE = PARTICIPATIONTYPE.LECTURER;
-
-    @Autowired
-    private ParticipationRepository participationRepository;
-
-    @Autowired
-    private ParticipationMapper participationMapper;
-
-    @Autowired
-    private ParticipationService participationService;
-
-    @Autowired
-    private EntityManager em;
-
-    @Autowired
-    private MockMvc restParticipationMockMvc;
-
-    private Participation participation;
-
-    /**
-     * Create an entity for this test.
-     *
-     * This is a static method, as tests for other entities might also need it,
-     * if they test an entity which requires the current entity.
-     */
-    public static Participation createEntity(EntityManager em) {
-        Participation participation = new Participation()
-            .uuid(DEFAULT_UUID)
-            .type(DEFAULT_TYPE);
-        return participation;
-    }
-    /**
-     * Create an updated entity for this test.
-     *
-     * This is a static method, as tests for other entities might also need it,
-     * if they test an entity which requires the current entity.
-     */
-    public static Participation createUpdatedEntity(EntityManager em) {
-        Participation participation = new Participation()
-            .uuid(UPDATED_UUID)
-            .type(UPDATED_TYPE);
-        return participation;
-    }
-
-    @BeforeEach
-    public void initTest() {
-        participation = createEntity(em);
-    }
-
-    @Test
-    @Transactional
-    public void createParticipation() throws Exception {
-        int databaseSizeBeforeCreate = participationRepository.findAll().size();
-        // Create the Participation
-        ParticipationDTO participationDTO = participationMapper.toDto(participation);
-        restParticipationMockMvc.perform(post("/api/participations").with(csrf())
-            .contentType(MediaType.APPLICATION_JSON)
-            .content(TestUtil.convertObjectToJsonBytes(participationDTO)))
-            .andExpect(status().isCreated());
-
-        // Validate the Participation in the database
-        List<Participation> participationList = participationRepository.findAll();
-        assertThat(participationList).hasSize(databaseSizeBeforeCreate + 1);
-        Participation testParticipation = participationList.get(participationList.size() - 1);
-        assertThat(testParticipation.getUuid()).isEqualTo(DEFAULT_UUID);
-        assertThat(testParticipation.getType()).isEqualTo(DEFAULT_TYPE);
-    }
-
-    @Test
-    @Transactional
-    public void createParticipationWithExistingId() throws Exception {
-        int databaseSizeBeforeCreate = participationRepository.findAll().size();
-
-        // Create the Participation with an existing ID
-        participation.setId(1L);
-        ParticipationDTO participationDTO = participationMapper.toDto(participation);
-
-        // An entity with an existing ID cannot be created, so this API call must fail
-        restParticipationMockMvc.perform(post("/api/participations").with(csrf())
-            .contentType(MediaType.APPLICATION_JSON)
-            .content(TestUtil.convertObjectToJsonBytes(participationDTO)))
-            .andExpect(status().isBadRequest());
-
-        // Validate the Participation in the database
-        List<Participation> participationList = participationRepository.findAll();
-        assertThat(participationList).hasSize(databaseSizeBeforeCreate);
-    }
-
-
-    @Test
-    @Transactional
-    public void getAllParticipations() throws Exception {
-        // Initialize the database
-        participationRepository.saveAndFlush(participation);
-
-        // Get all the participationList
-        restParticipationMockMvc.perform(get("/api/participations?sort=id,desc"))
-            .andExpect(status().isOk())
-            .andExpect(content().contentType(MediaType.APPLICATION_JSON_VALUE))
-            .andExpect(jsonPath("$.[*].id").value(hasItem(participation.getId().intValue())))
-            .andExpect(jsonPath("$.[*].uuid").value(hasItem(DEFAULT_UUID.toString())))
-            .andExpect(jsonPath("$.[*].type").value(hasItem(DEFAULT_TYPE.toString())));
-    }
-    
-    @Test
-    @Transactional
-    public void getParticipation() throws Exception {
-        // Initialize the database
-        participationRepository.saveAndFlush(participation);
-
-        // Get the participation
-        restParticipationMockMvc.perform(get("/api/participations/{id}", participation.getId()))
-            .andExpect(status().isOk())
-            .andExpect(content().contentType(MediaType.APPLICATION_JSON_VALUE))
-            .andExpect(jsonPath("$.id").value(participation.getId().intValue()))
-            .andExpect(jsonPath("$.uuid").value(DEFAULT_UUID.toString()))
-            .andExpect(jsonPath("$.type").value(DEFAULT_TYPE.toString()));
-    }
-    @Test
-    @Transactional
-    public void getNonExistingParticipation() throws Exception {
-        // Get the participation
-        restParticipationMockMvc.perform(get("/api/participations/{id}", Long.MAX_VALUE))
-            .andExpect(status().isNotFound());
-    }
-
-    @Test
-    @Transactional
-    public void updateParticipation() throws Exception {
-        // Initialize the database
-        participationRepository.saveAndFlush(participation);
-
-        int databaseSizeBeforeUpdate = participationRepository.findAll().size();
-
-        // Update the participation
-        Participation updatedParticipation = participationRepository.findById(participation.getId()).get();
-        // Disconnect from session so that the updates on updatedParticipation are not directly saved in db
-        em.detach(updatedParticipation);
-        updatedParticipation
-            .uuid(UPDATED_UUID)
-            .type(UPDATED_TYPE);
-        ParticipationDTO participationDTO = participationMapper.toDto(updatedParticipation);
-
-        restParticipationMockMvc.perform(put("/api/participations").with(csrf())
-            .contentType(MediaType.APPLICATION_JSON)
-            .content(TestUtil.convertObjectToJsonBytes(participationDTO)))
-            .andExpect(status().isOk());
-
-        // Validate the Participation in the database
-        List<Participation> participationList = participationRepository.findAll();
-        assertThat(participationList).hasSize(databaseSizeBeforeUpdate);
-        Participation testParticipation = participationList.get(participationList.size() - 1);
-        assertThat(testParticipation.getUuid()).isEqualTo(UPDATED_UUID);
-        assertThat(testParticipation.getType()).isEqualTo(UPDATED_TYPE);
-    }
-
-    @Test
-    @Transactional
-    public void updateNonExistingParticipation() throws Exception {
-        int databaseSizeBeforeUpdate = participationRepository.findAll().size();
-
-        // Create the Participation
-        ParticipationDTO participationDTO = participationMapper.toDto(participation);
-
-        // If the entity doesn't have an ID, it will throw BadRequestAlertException
-        restParticipationMockMvc.perform(put("/api/participations").with(csrf())
-            .contentType(MediaType.APPLICATION_JSON)
-            .content(TestUtil.convertObjectToJsonBytes(participationDTO)))
-            .andExpect(status().isBadRequest());
-
-        // Validate the Participation in the database
-        List<Participation> participationList = participationRepository.findAll();
-        assertThat(participationList).hasSize(databaseSizeBeforeUpdate);
-    }
-
-    @Test
-    @Transactional
-    public void deleteParticipation() throws Exception {
-        // Initialize the database
-        participationRepository.saveAndFlush(participation);
-
-        int databaseSizeBeforeDelete = participationRepository.findAll().size();
-
-        // Delete the participation
-        restParticipationMockMvc.perform(delete("/api/participations/{id}", participation.getId()).with(csrf())
-            .accept(MediaType.APPLICATION_JSON))
-            .andExpect(status().isNoContent());
-
-        // Validate the database contains one less item
-        List<Participation> participationList = participationRepository.findAll();
-        assertThat(participationList).hasSize(databaseSizeBeforeDelete - 1);
-    }
-}
