+++ src/test/java/de/uni_stuttgart/it_rex/course/web/rest/written/CourseResourceIT.java
@@ -1,0 +1,285 @@
+package de.uni_stuttgart.it_rex.course.web.rest.written;
+
+import de.uni_stuttgart.it_rex.course.CourseServiceApp;
+import de.uni_stuttgart.it_rex.course.config.TestSecurityConfiguration;
+import de.uni_stuttgart.it_rex.course.domain.written.Course;
+import de.uni_stuttgart.it_rex.course.repository.written.CourseRepository;
+import de.uni_stuttgart.it_rex.course.service.written.CourseService;
+import de.uni_stuttgart.it_rex.course.service.written.dto.CourseDTO;
+import de.uni_stuttgart.it_rex.course.service.written.mapper.CourseMapper;
+
+import de.uni_stuttgart.it_rex.course.web.rest.TestUtil;
+import de.uni_stuttgart.it_rex.course.web.rest.written.CourseResource;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.http.MediaType;
+import org.springframework.security.test.context.support.WithMockUser;
+import org.springframework.test.web.servlet.MockMvc;
+import org.springframework.transaction.annotation.Transactional;
+
+import javax.persistence.EntityManager;
+import java.time.LocalDate;
+import java.time.ZoneId;
+import java.util.List;
+import java.util.UUID;
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.hamcrest.Matchers.hasItem;
+import static org.springframework.security.test.web.servlet.request.SecurityMockMvcRequestPostProcessors.csrf;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
+
+import de.uni_stuttgart.it_rex.course.domain.enumeration.PUBLISHSTATE;
+/**
+ * Integration tests for the {@link CourseResource} REST controller.
+ */
+@SpringBootTest(classes = { CourseServiceApp.class, TestSecurityConfiguration.class })
+@AutoConfigureMockMvc
+@WithMockUser
+public class CourseResourceIT {
+
+    private static final UUID DEFAULT_UUID = UUID.randomUUID();
+    private static final UUID UPDATED_UUID = UUID.randomUUID();
+
+    private static final String DEFAULT_NAME = "AAAAAAAAAA";
+    private static final String UPDATED_NAME = "BBBBBBBBBB";
+
+    private static final LocalDate DEFAULT_START_DATE = LocalDate.ofEpochDay(0L);
+    private static final LocalDate UPDATED_START_DATE = LocalDate.now(ZoneId.systemDefault());
+
+    private static final LocalDate DEFAULT_END_DATE = LocalDate.ofEpochDay(0L);
+    private static final LocalDate UPDATED_END_DATE = LocalDate.now(ZoneId.systemDefault());
+
+    private static final Integer DEFAULT_MAX_FOOD_SUM = 1;
+    private static final Integer UPDATED_MAX_FOOD_SUM = 2;
+
+    private static final String DEFAULT_COURSE_DESCRIPTION = "AAAAAAAAAA";
+    private static final String UPDATED_COURSE_DESCRIPTION = "BBBBBBBBBB";
+
+    private static final PUBLISHSTATE DEFAULT_PUBLISH_STATE = PUBLISHSTATE.UNPUBLISHED;
+    private static final PUBLISHSTATE UPDATED_PUBLISH_STATE = PUBLISHSTATE.PUBLISHED;
+
+    @Autowired
+    private CourseRepository courseRepository;
+
+    @Autowired
+    private CourseMapper courseMapper;
+
+    @Autowired
+    private CourseService courseService;
+
+    @Autowired
+    private EntityManager em;
+
+    @Autowired
+    private MockMvc restCourseMockMvc;
+
+    private Course course;
+
+    /**
+     * Create an entity for this test.
+     *
+     * This is a static method, as tests for other entities might also need it,
+     * if they test an entity which requires the current entity.
+     */
+    public static Course createEntity(EntityManager em) {
+        Course course = new Course()
+            .name(DEFAULT_NAME)
+            .startDate(DEFAULT_START_DATE)
+            .endDate(DEFAULT_END_DATE)
+            .maxFoodSum(DEFAULT_MAX_FOOD_SUM)
+            .courseDescription(DEFAULT_COURSE_DESCRIPTION)
+            .publishState(DEFAULT_PUBLISH_STATE);
+        return course;
+    }
+    /**
+     * Create an updated entity for this test.
+     *
+     * This is a static method, as tests for other entities might also need it,
+     * if they test an entity which requires the current entity.
+     */
+    public static Course createUpdatedEntity(EntityManager em) {
+        Course course = new Course()
+            .name(UPDATED_NAME)
+            .startDate(UPDATED_START_DATE)
+            .endDate(UPDATED_END_DATE)
+            .maxFoodSum(UPDATED_MAX_FOOD_SUM)
+            .courseDescription(UPDATED_COURSE_DESCRIPTION)
+            .publishState(UPDATED_PUBLISH_STATE);
+        return course;
+    }
+
+    @BeforeEach
+    public void initTest() {
+        course = createEntity(em);
+    }
+
+    @Test
+    @Transactional
+    public void createCourse() throws Exception {
+        int databaseSizeBeforeCreate = courseRepository.findAll().size();
+        // Create the Course
+        CourseDTO courseDTO = courseMapper.toDto(course);
+        restCourseMockMvc.perform(post("/api/courses").with(csrf())
+            .contentType(MediaType.APPLICATION_JSON)
+            .content(TestUtil.convertObjectToJsonBytes(courseDTO)))
+            .andExpect(status().isCreated());
+
+        // Validate the Course in the database
+        List<Course> courseList = courseRepository.findAll();
+        assertThat(courseList).hasSize(databaseSizeBeforeCreate + 1);
+        Course testCourse = courseList.get(courseList.size() - 1);
+        assertThat(testCourse.getName()).isEqualTo(DEFAULT_NAME);
+        assertThat(testCourse.getStartDate()).isEqualTo(DEFAULT_START_DATE);
+        assertThat(testCourse.getEndDate()).isEqualTo(DEFAULT_END_DATE);
+        assertThat(testCourse.getMaxFoodSum()).isEqualTo(DEFAULT_MAX_FOOD_SUM);
+        assertThat(testCourse.getCourseDescription()).isEqualTo(DEFAULT_COURSE_DESCRIPTION);
+        assertThat(testCourse.getPublishState()).isEqualTo(DEFAULT_PUBLISH_STATE);
+    }
+
+    @Test
+    @Transactional
+    public void createCourseWithExistingId() throws Exception {
+        int databaseSizeBeforeCreate = courseRepository.findAll().size();
+
+        // Create the Course with an existing ID
+        course.setId(UUID.randomUUID());
+        CourseDTO courseDTO = courseMapper.toDto(course);
+
+        // An entity with an existing ID cannot be created, so this API call must fail
+        restCourseMockMvc.perform(post("/api/courses").with(csrf())
+            .contentType(MediaType.APPLICATION_JSON)
+            .content(TestUtil.convertObjectToJsonBytes(courseDTO)))
+            .andExpect(status().isBadRequest());
+
+        // Validate the Course in the database
+        List<Course> courseList = courseRepository.findAll();
+        assertThat(courseList).hasSize(databaseSizeBeforeCreate);
+    }
+
+
+    @Test
+    @Transactional
+    public void getAllCourses() throws Exception {
+        // Initialize the database
+        courseRepository.saveAndFlush(course);
+
+        // Get all the courseList
+        restCourseMockMvc.perform(get("/api/courses?sort=id,desc"))
+            .andExpect(status().isOk())
+            .andExpect(content().contentType(MediaType.APPLICATION_JSON_VALUE))
+            .andExpect(jsonPath("$.[*].id").value(hasItem(course.getId().toString())))
+            .andExpect(jsonPath("$.[*].uuid").value(hasItem(DEFAULT_UUID.toString())))
+            .andExpect(jsonPath("$.[*].name").value(hasItem(DEFAULT_NAME)))
+            .andExpect(jsonPath("$.[*].startDate").value(hasItem(DEFAULT_START_DATE.toString())))
+            .andExpect(jsonPath("$.[*].endDate").value(hasItem(DEFAULT_END_DATE.toString())))
+            .andExpect(jsonPath("$.[*].maxFoodSum").value(hasItem(DEFAULT_MAX_FOOD_SUM)))
+            .andExpect(jsonPath("$.[*].courseDescription").value(hasItem(DEFAULT_COURSE_DESCRIPTION.toString())))
+            .andExpect(jsonPath("$.[*].publishState").value(hasItem(DEFAULT_PUBLISH_STATE.toString())));
+    }
+    
+    @Test
+    @Transactional
+    public void getCourse() throws Exception {
+        // Initialize the database
+        courseRepository.saveAndFlush(course);
+
+        // Get the course
+        restCourseMockMvc.perform(get("/api/courses/{id}", course.getId()))
+            .andExpect(status().isOk())
+            .andExpect(content().contentType(MediaType.APPLICATION_JSON_VALUE))
+            .andExpect(jsonPath("$.id").value(course.getId().toString()))
+            .andExpect(jsonPath("$.uuid").value(DEFAULT_UUID.toString()))
+            .andExpect(jsonPath("$.name").value(DEFAULT_NAME))
+            .andExpect(jsonPath("$.startDate").value(DEFAULT_START_DATE.toString()))
+            .andExpect(jsonPath("$.endDate").value(DEFAULT_END_DATE.toString()))
+            .andExpect(jsonPath("$.maxFoodSum").value(DEFAULT_MAX_FOOD_SUM))
+            .andExpect(jsonPath("$.courseDescription").value(DEFAULT_COURSE_DESCRIPTION.toString()))
+            .andExpect(jsonPath("$.publishState").value(DEFAULT_PUBLISH_STATE.toString()));
+    }
+    @Test
+    @Transactional
+    public void getNonExistingCourse() throws Exception {
+        // Get the course
+        restCourseMockMvc.perform(get("/api/courses/{id}", Long.MAX_VALUE))
+            .andExpect(status().isNotFound());
+    }
+
+    @Test
+    @Transactional
+    public void updateCourse() throws Exception {
+        // Initialize the database
+        courseRepository.saveAndFlush(course);
+
+        int databaseSizeBeforeUpdate = courseRepository.findAll().size();
+
+        // Update the course
+        Course updatedCourse = courseRepository.findById(course.getId()).get();
+        // Disconnect from session so that the updates on updatedCourse are not directly saved in db
+        em.detach(updatedCourse);
+        updatedCourse
+            .name(UPDATED_NAME)
+            .startDate(UPDATED_START_DATE)
+            .endDate(UPDATED_END_DATE)
+            .maxFoodSum(UPDATED_MAX_FOOD_SUM)
+            .courseDescription(UPDATED_COURSE_DESCRIPTION)
+            .publishState(UPDATED_PUBLISH_STATE);
+        CourseDTO courseDTO = courseMapper.toDto(updatedCourse);
+
+        restCourseMockMvc.perform(put("/api/courses").with(csrf())
+            .contentType(MediaType.APPLICATION_JSON)
+            .content(TestUtil.convertObjectToJsonBytes(courseDTO)))
+            .andExpect(status().isOk());
+
+        // Validate the Course in the database
+        List<Course> courseList = courseRepository.findAll();
+        assertThat(courseList).hasSize(databaseSizeBeforeUpdate);
+        Course testCourse = courseList.get(courseList.size() - 1);
+        assertThat(testCourse.getName()).isEqualTo(UPDATED_NAME);
+        assertThat(testCourse.getStartDate()).isEqualTo(UPDATED_START_DATE);
+        assertThat(testCourse.getEndDate()).isEqualTo(UPDATED_END_DATE);
+        assertThat(testCourse.getMaxFoodSum()).isEqualTo(UPDATED_MAX_FOOD_SUM);
+        assertThat(testCourse.getCourseDescription()).isEqualTo(UPDATED_COURSE_DESCRIPTION);
+        assertThat(testCourse.getPublishState()).isEqualTo(UPDATED_PUBLISH_STATE);
+    }
+
+    @Test
+    @Transactional
+    public void updateNonExistingCourse() throws Exception {
+        int databaseSizeBeforeUpdate = courseRepository.findAll().size();
+
+        // Create the Course
+        CourseDTO courseDTO = courseMapper.toDto(course);
+
+        // If the entity doesn't have an ID, it will throw BadRequestAlertException
+        restCourseMockMvc.perform(put("/api/courses").with(csrf())
+            .contentType(MediaType.APPLICATION_JSON)
+            .content(TestUtil.convertObjectToJsonBytes(courseDTO)))
+            .andExpect(status().isBadRequest());
+
+        // Validate the Course in the database
+        List<Course> courseList = courseRepository.findAll();
+        assertThat(courseList).hasSize(databaseSizeBeforeUpdate);
+    }
+
+    @Test
+    @Transactional
+    public void deleteCourse() throws Exception {
+        // Initialize the database
+        courseRepository.saveAndFlush(course);
+
+        int databaseSizeBeforeDelete = courseRepository.findAll().size();
+
+        // Delete the course
+        restCourseMockMvc.perform(delete("/api/courses/{id}", course.getId()).with(csrf())
+            .accept(MediaType.APPLICATION_JSON))
+            .andExpect(status().isNoContent());
+
+        // Validate the database contains one less item
+        List<Course> courseList = courseRepository.findAll();
+        assertThat(courseList).hasSize(databaseSizeBeforeDelete - 1);
+    }
+}
